Output buffers fixed

+++ Altium/FileSorter/Models/BatchDataForSort.cs
@@ -2,7 +2,7 @@
 {
 	internal struct BatchDataForSort
 	{
-		public BatchDataForSort(bool isItFirstFile, bool isItLastFile, string nameWithPath, int startOffset, int interruptionOffset)
+		public BatchDataForSort(bool isItFirstFile, bool isItLastFile, string nameWithPath, long startOffset, long interruptionOffset)
 		{
 			IsItFirstFile = isItFirstFile;
 			IsItLastFile = isItLastFile;
@@ -14,7 +14,7 @@
 		public bool IsItFirstFile { get; }
 		public bool IsItLastFile { get; }
 		public string NameWithPath { get; }
-		public int StartOffset { get; }
-		public int InterruptionOffset { get; }
+		public long StartOffset { get; }
+		public long InterruptionOffset { get; }
 	}
 }
